Feature/pundit role authorization

+++ app/controllers/invitations_controller.rb
@@ -7,12 +7,14 @@ class InvitationsController < ApplicationController
   around_action :with_current_account_tenant
 
   def new
-    @invitation = Invitation.new
+    @invitation = current_user.account.invitations.build
+    authorize @invitation
   end
 
   def create
     @invitation = current_user.account.invitations.build(invitation_params)
     @invitation.invited_by = current_user
+    authorize @invitation
 
     if @invitation.save
       InvitationMailer.invite(@invitation).deliver_now
@@ -24,6 +26,7 @@ class InvitationsController < ApplicationController
 
   def destroy
     invitation = current_user.account.invitations.pending.find(params[:id])
+    authorize invitation
     invitation.destroy!
     redirect_to dashboard_path, notice: t(".notice")
   end
